# pref(system): Clean up overdesign and reduce size