## Before

esm/index.mjs  15.3kb

clk: ~0.05 GHz
cpu: Apple M3 Max
runtime: node 22.9.0 (arm64-darwin)

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `790.38 ns/iter` | `773.45 ns` | `800.21 ns` | `820.24 ns` | `825.65 ns` |
| propagate: 10 * 1    | `  7.03 µs/iter` | `  6.92 µs` | `  7.09 µs` | `  7.14 µs` | `  7.14 µs` |
| propagate: 100 * 1   | ` 68.82 µs/iter` | ` 66.13 µs` | ` 67.50 µs` | ` 90.67 µs` | ` 94.79 µs` |
| propagate: 1 * 10    | `  4.02 µs/iter` | `  4.00 µs` | `  4.02 µs` | `  4.06 µs` | `  4.13 µs` |
| propagate: 10 * 10   | ` 39.42 µs/iter` | ` 38.80 µs` | ` 39.67 µs` | ` 39.91 µs` | ` 40.14 µs` |
| propagate: 100 * 10  | `399.07 µs/iter` | `381.33 µs` | `408.83 µs` | `449.50 µs` | `468.04 µs` |
| propagate: 1 * 100   | ` 35.48 µs/iter` | ` 35.00 µs` | ` 35.60 µs` | ` 35.72 µs` | ` 35.84 µs` |
| propagate: 10 * 100  | `355.62 µs/iter` | `347.46 µs` | `355.63 µs` | `386.71 µs` | `404.17 µs` |
| propagate: 100 * 100 | `  3.55 ms/iter` | `  3.50 ms` | `  3.56 ms` | `  3.64 ms` | `  3.66 ms` |

## After

esm/index.mjs  15.0kb

clk: ~0.05 GHz
cpu: Apple M3 Max
runtime: node 22.9.0 (arm64-darwin)

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `786.68 ns/iter` | `770.51 ns` | `792.65 ns` | `807.85 ns` | `814.28 ns` |
| propagate: 10 * 1    | `  6.87 µs/iter` | `  6.79 µs` | `  6.90 µs` | `  6.94 µs` | `  6.95 µs` |
| propagate: 100 * 1   | ` 67.66 µs/iter` | ` 65.00 µs` | ` 66.25 µs` | ` 89.46 µs` | `100.38 µs` |
| propagate: 1 * 10    | `  3.99 µs/iter` | `  3.96 µs` | `  3.99 µs` | `  4.05 µs` | `  4.05 µs` |
| propagate: 10 * 10   | ` 39.59 µs/iter` | ` 38.96 µs` | ` 39.91 µs` | ` 39.98 µs` | ` 40.06 µs` |
| propagate: 100 * 10  | `394.43 µs/iter` | `380.50 µs` | `400.00 µs` | `439.25 µs` | `451.88 µs` |
| propagate: 1 * 100   | ` 35.38 µs/iter` | ` 35.01 µs` | ` 35.63 µs` | ` 35.72 µs` | ` 35.85 µs` |
| propagate: 10 * 100  | `356.91 µs/iter` | `349.58 µs` | `357.29 µs` | `391.96 µs` | `408.08 µs` |
| propagate: 100 * 100 | `  3.52 ms/iter` | `  3.47 ms` | `  3.54 ms` | `  3.63 ms` | `  3.63 ms` |

## Why?

Reusing Link objects does not save or speed up, but leaves many design traps. Linking empty objects all the time will result in garbage collection not being possible, and the next stage will reset the link before releasing it.

This PR removes this excessive design and directly discards the link when it is no longer needed to facilitate timely garbage collection.