@medz 2024-12-17
In addition, I found that bench allows some data differences each time, but it does not affect the final result. It may be that there is a discrepancy of 1*1.