@johnsoncodehk 2024-12-17
Thank you for your PR. Our bench script cannot accurately test the impact of link pool on performance. In the bench test case, clearTrack is never triggered, so the link pool is never used. The performance improvement is due to the fact that there is no need to perform `linkNewDep!= = undefined` judgment in linkNewDep.

Secondly, because the bench script uses `--jitless`, it can only be used as a reference, but it cannot correctly reflect the performance of the code in the real node app. We need to test it in https://github.com/transitive-bullshit/js-reactivity-benchmark to determine the final performance impact. (Using link pool should be significantly improved)