@medz 2024-12-17
- A：The PR
- B: 0.4.4

| test| A avg | B avg | diff(B vs A) |
|--------|----------|----------|--------------|
| avoidablePropagation | 188.04 | 191.34 | +1.75% |
| broadPropagation | 115.67 | 114.23 | -1.24% |
| deepPropagation | 53.74 | 53.73 | -0.02% |
| diamond | 150.43 | 142.07 | -5.56% ✓ |
| mux | 144.05 | 144.85 | +0.56% |
| repeatedObservers | 19.03 | 18.65 | -2.00% |
| triangle | 50.16 | 50.43 | +0.54% |
| unstable | 16.50 | 15.95 | -3.33% ✓ |
| molBench | 285.34 | 286.09 | +0.26% |
| createDataSignals | 8.83 | 7.71 | -12.68% ✓ |
| createComputations1to2 | 4.08 | 2.13 | -47.79% ✓ |
| createComputations1to4 | 2.84 | 2.86 | +0.70% |
| updateComputations1to1000 | 1.49 | 1.32 | -11.41% ✓ |
| 10x5 (simple component) | 99.60 | 102.28 | +2.69% |
| 10x10 (dynamic component) | 111.94 | 107.63 | -3.85% ✓ |
| 1000x12 (large web app) | 181.22 | 183.08 | +1.03% |
| 1000x5 (wide dense) | 256.05 | 261.30 | +2.05% |
| 5x500 (deep) | 100.56 | 104.59 | +4.01% |

I think this PR can be closed.

From the test results, PR (A) has better performance in some aspects (but the difference is small), but overall 0.4.4 (B) is obviously better.

> Can you help me solve this puzzle? I noticed that v8 is much more resource-intensive to create Object than to reuse it, which is completely different from the Rust and Dart languages ​​I often use.
>
> I am porting the Dart version, and in Dart, directly discarding resources is much faster than clearing properties and filling them again. (Maybe it has something to do with the memory allocation strategy)